+++ src/core/__init__.py
@@ -8,8 +8,8 @@
 
 # Re-export main components from submodules for convenience
 from .env_server import *
+from .client_types import StepResult
 from .http_env_client import HTTPEnvClient
-from .types import StepResult
 
 # Note: MCP module doesn't export anything yet
