+++ src/envs/atari_env/client.py
@@ -15,8 +15,9 @@ from __future__ import annotations
 
 from typing import Any, Dict, TYPE_CHECKING
 
+from core.client_types import StepResult
+
 from core.http_env_client import HTTPEnvClient
-from core.types import StepResult
 
 from .models import AtariAction, AtariObservation, AtariState
