+++ src/envs/finrl_env/client.py
@@ -13,9 +13,10 @@ over HTTP.
 
 from typing import Any, Dict
 
+from core.client_types import StepResult
+
 from core.env_server.types import State
 from core.http_env_client import HTTPEnvClient
-from core.types import StepResult
 
 from .models import FinRLAction, FinRLObservation
