+++ src/envs/sumo_rl_env/client.py
@@ -13,8 +13,9 @@ control environment over HTTP.
 
 from typing import Any, Dict
 
+from core.client_types import StepResult
+
 from core.http_env_client import HTTPEnvClient
-from core.types import StepResult
 
 from .models import SumoAction, SumoObservation, SumoState
